produkt

+++ src/ModelLayer/MassProduct.java
@@ -4,7 +4,7 @@ public class MassProduct extends Product{
     /**
      * constructor til objekter af class MassProduct
      */
-    public MassProduct(int barcode, String name, String description, int price, int amount, int min, int max){
+    public MassProduct(int barcode, String name, String description, double price, int amount, int min, int max){
         super(barcode, name, description, price, amount, min, max);
     }
